added support for registrations endpoint
Some OAuth providers provide an endpoint for a registration page instead of a login page. By configuring the `registrations_path` and `registrations_endpoint`, the `openidc_authorize` function will redirect to the configured endpoint including the same parameters as the authorization endpoint and prepare the session.

Can be used e.g. for providing direct Links to Keycloak's registration page.